# GestionParqueadero/LogicaNegocio/Modelos/VehiculoModel.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace LogicaNegocio.Modelos
{
    public class VehiculoModel
    {
        public string PlacaVehiculo { get; set; }
        public string? MarcaVehiculo { get; set; }
        public string? ModeloVehiculo { get; set; }
        public string? ColorVehiculo { get; set; }
        public string? TipoVehiculo { get; set; }
    }
}