# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebApiGestionParqueaderos",
    "version": "1.0"
  },
  "paths": {
    "/api/Caja/getCajas": {
      "get": {
        "tags": [
          "Caja"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CajaModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CajaModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CajaModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Caja/getCaja": {
      "get": {
        "tags": [
          "Caja"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CajaModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CajaModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CajaModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Caja/postCaja": {
      "post": {
        "tags": [
          "Caja"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CajaModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CajaModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CajaModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Caja/deleteCaja": {
      "delete": {
        "tags": [
          "Caja"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/getClientes": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Cliente/getCliente": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Cliente/postCliente": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/deleteCliente": {
      "delete": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Login/Authorize": {
      "post": {
        "tags": [
          "Login"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Login": {
      "get": {
        "tags": [
          "Login"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Registro/getRegistros": {
      "get": {
        "tags": [
          "Registro"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistroModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistroModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistroModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Registro/getRegistro": {
      "get": {
        "tags": [
          "Registro"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegistroModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistroModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistroModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Registro/postRegistro": {
      "post": {
        "tags": [
          "Registro"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Registro/deleteRegistro": {
      "delete": {
        "tags": [
          "Registro"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Vehiculo/getVehiculos": {
      "get": {
        "tags": [
          "Vehiculo"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehiculoModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehiculoModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehiculoModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Vehiculo/getVehiculo": {
      "get": {
        "tags": [
          "Vehiculo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Vehiculo/postVehiculos": {
      "post": {
        "tags": [
          "Vehiculo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Vehiculo/deleteVehiculo": {
      "delete": {
        "tags": [
          "Vehiculo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CajaModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idRegistro": {
            "type": "integer",
            "format": "int64"
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "valor": {
            "type": "number",
            "format": "float"
          },
          "idCliente": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClienteModel": {
        "type": "object",
        "properties": {
          "numeroDocumento": {
            "type": "string",
            "nullable": true
          },
          "nombres": {
            "pattern": "\\A[^\\d_]+\\z",
            "type": "string",
            "nullable": true
          },
          "apellidos": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegistroModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "fechaIngreso": {
            "type": "string",
            "format": "date-time"
          },
          "fechaSalida": {
            "type": "string",
            "format": "date-time"
          },
          "placaVehiculo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UsuarioModel": {
        "required": [
          "clave",
          "nombreUsuario"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "nombreUsuario": {
            "type": "string"
          },
          "clave": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "VehiculoModel": {
        "type": "object",
        "properties": {
          "placaVehiculo": {
            "type": "string",
            "nullable": true
          },
          "marcaVehiculo": {
            "type": "string",
            "nullable": true
          },
          "modeloVehiculo": {
            "type": "string",
            "nullable": true
          },
          "colorVehiculo": {
            "type": "string",
            "nullable": true
          },
          "tipoVehiculo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}